--- a/eslint.config.js
+++ b/eslint.config.js
@@ -2,6 +2,7 @@ import js from "@eslint/js";
 import nodePlugin from "eslint-plugin-node";
 import jestPlugin from "eslint-plugin-jest";
 import tsParser from "@typescript-eslint/parser";
+import tsPlugin from "@typescript-eslint/eslint-plugin";
 import prettierConfig from "eslint-config-prettier";
 
 export default [
@@ -41,7 +42,7 @@ export default [
           caughtErrorsIgnorePattern: "^_",
         },
       ],
-      "no-console": "off",
+      "no-console": "error",
       semi: ["error", "always"],
       quotes: ["error", "double"],
       "comma-dangle": ["error", "always-multiline"],
@@ -62,6 +63,9 @@ export default [
   },
   {
     files: ["**/*.ts"],
+    plugins: {
+      "@typescript-eslint": tsPlugin,
+    },
     rules: {
       "no-unused-vars": [
         "error",
@@ -74,10 +78,20 @@ export default [
       "no-undef": "off", // TypeScript handles this
       "no-case-declarations": "off",
       "no-useless-escape": "off",
+      // TypeScript specific rules
+      "@typescript-eslint/no-explicit-any": "error",
+      "@typescript-eslint/no-unused-vars": [
+        "error", 
+        {
+          args: "none",
+          varsIgnorePattern: "^_",
+          caughtErrorsIgnorePattern: "^_",
+        }
+      ],
     },
   },
   {
-    files: ["**/*.test.js", "**/*.spec.js", "tests/**/*.js"],
+    files: ["**/*.test.js", "**/*.spec.js", "tests/**/*.js", "**/*.test.ts", "**/*.spec.ts", "**/__tests__/**/*.ts"],
     plugins: {
       jest: jestPlugin,
     },
@@ -100,6 +114,14 @@ export default [
       "jest/no-identical-title": "error",
       "jest/prefer-to-have-length": "warn",
       "jest/valid-expect": "error",
+      "no-console": "off", // Allow console in tests
+      "@typescript-eslint/no-explicit-any": "off", // Allow any in tests
+    },
+  },
+  {
+    files: ["src/utils/logger.ts", "scripts/**/*.js", "scripts/**/*.cjs"],
+    rules: {
+      "no-console": "off", // Logger utility and scripts need console access
     },
   },
   {