… .c8rc.json
… .github/workflows/coverage-guard.yml
… .nycrc.json
… CLAUDE.md
… COVERAGE_STRATEGY.md
… PRODUCTION.md
… README.md
… SECURITY-AUDIT.md
… TESTING_GUIDELINES.md
… docs/BADGE_UPDATES.md
… docs/INCREMENTAL_COVERAGE.md
… jest.typescript.config.json
… package.json
… scripts/fix-coverage-reporting.cjs
… scripts/generate-coverage-report.cjs
… scripts/performance-benchmarks.cjs
… src/cache/CacheManager.ts
… src/cache/HttpCacheWrapper.ts
… src/cache/__tests__/HttpCacheWrapper.test.ts
… src/client/CachedWordPressClient.ts
… src/client/MockWordPressClient.ts
… src/client/api.ts
… src/client/auth.ts
… src/client/managers/AuthenticationManager.ts
… src/client/managers/BaseManager.ts
… src/client/managers/RequestManager.ts
… src/config/Config.ts
… src/config/ServerConfiguration.ts
… src/index.ts
… src/performance/MetricsCollector.ts